+++ src/main/java/org/archive/io/warc/WARCWriter.java
@@ -37,7 +37,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import org.apache.commons.lang.StringUtils;
-import org.archive.io.ArchiveFileConstants;
+import org.archive.format.ArchiveFileConstants;
 import org.archive.io.UTF8Bytes;
 import org.archive.io.WriterPoolMember;
 import org.archive.util.ArchiveUtils;
@@ -343,9 +343,9 @@ implements WARCConstants {
         recordInfo.setMimetype("application/warc-fields");
 
         // Strip .open suffix if present.
-        if (filename.endsWith(WriterPoolMember.OCCUPIED_SUFFIX)) {
+        if (filename.endsWith(ArchiveFileConstants.OCCUPIED_SUFFIX)) {
         	filename = filename.substring(0,
-        		filename.length() - WriterPoolMember.OCCUPIED_SUFFIX.length());
+        		filename.length() - ArchiveFileConstants.OCCUPIED_SUFFIX.length());
         }
         recordInfo.addExtraHeader(HEADER_KEY_FILENAME, filename);
         if (description != null && description.length() > 0) {